[UPD] Force the about dialog to display the version 0.1a4 (Mokka is still in alpha). [FIX] The combobox for the visual configurations has no frame under Windows XP. [FIX] Two cl warnings.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@378 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Tools/Mokka/Viz3DWidget.cpp
@@ -421,7 +421,7 @@ void Viz3DWidget::updateDisplayedMarkersList(vtkObject* caller, unsigned long /*
     return;
     
   QVector<int> ids = QVector<int>(indexes->GetSize());
-  for (size_t i = 0; i < ids.size() ; ++i)
+  for (int i = 0; i < ids.size() ; ++i)
     ids[i] = indexes->GetValue(i);
   emit visibleMarkersChanged(ids);
 };
